delete unused files and move all samples into sample dir
