+++ sample/CEMAgent.py
@@ -1,0 +1,45 @@
+import numpy as np
+from gym_core import tgym
+from rl.agents import cem
+
+from keras.models import Sequential
+from keras.layers import Dense, Activation, Flatten
+from keras.optimizers import Adam
+
+from rl.agents.cem import CEMAgent
+from rl.memory import EpisodeParameterMemory
+
+
+# Get the environment and extract the number of actions.
+env = tgym.TradingGymEnv(episode_type=0)
+np.random.seed(123)
+env.seed(123)
+
+nb_actions = env.action_space.n
+obs_dim = env.observation_space.shape[0]
+
+model = Sequential()
+model.add(Flatten(input_shape=(1,) + env.observation_space.shape))
+model.add(Dense(nb_actions))
+model.add(Activation('softmax'))
+
+print(model.summary())
+
+# Finally, we configure and compile our agent. You can use every built-in Keras optimizer and
+# even the metrics!
+memory = EpisodeParameterMemory(limit=1000, window_length=1)
+
+cem = CEMAgent(model=model, nb_actions=nb_actions, memory=memory,
+               batch_size=50, nb_steps_warmup=2000, train_interval=50, elite_frac=0.05)
+cem.compile()
+
+# Okay, now it's time to learn something! We visualiz   e the training here for show, but this
+# slows down training quite a lot. You can always safely abort the training prematurely using
+# Ctrl + C.
+cem.fit(env, nb_steps=100000, visualize=False, verbose=2)
+
+# After training is done, we save the best weights.
+cem.save_weights('cem_{}_params.h5f'.format(ENV_NAME), overwrite=True)
+
+# Finally, evaluate our algorithm for 5 episodes.
+cem.test(env, nb_episodes=5, visualize=True)
