--- sample/test-debug.py
@@ -1,11 +1,0 @@
-
-import numpy as np
-
-x = np.array([1,2])
-
-
-for a in [1,2,3,4] :
-    for b in [1, 2, 3, 4]:
-        print(a,b)
-
-print(x)
